Refactor user data fetching to be centralised
