+++ client/src/contexts/UserContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const UserContext = createContext();
+
+export default UserContext;
